requirements.txt
discord.py==2.4.0
PyQt5==5.15.11
PyQt5_sip==12.15.0
python-dotenv==1.0.1
